Merge current development branch to stable

+++ src/assert.cpp
@@ -5,8 +5,9 @@
 
 OSL_BEGIN_NAMESPACE
 
-void assert_failed(const char *file, u64 line, const char* message) {
-    fprintf(stderr, "%s:%lu: assertion failed:\n\t%s\n\t", file, line, message);
+void assert_failed(const char *file, u32 line, const char* message)
+{
+    fprintf(stderr, "%s:%u: assertion failed:\n\t%s\n\t", file, line, message);
     exit(1);
 }
